+++ lib.php
@@ -56,6 +56,18 @@ class format_columns extends format_base {
     }
 
     /**
+     * Returns the default section name for the format.
+     *
+     * @param stdClass $section Section object from database or just field course_sections section
+     * @return string The default value for the section name.
+     */
+    public function get_default_section_name($section) {
+        /* Follow the same logic so that this method is supported.  The MDL-51610 enchancement refactored things,
+           but that is not appropriate for us. */
+        return $this->get_section_name($section);
+    }
+
+    /**
      * Gets the name for the provided section.
      *
      * @param stdClass $section The section.