+++ Changes.md
@@ -4,6 +4,7 @@ Version 3.0.1
   1. First version for Moodle 3.0.
   2. Added TravisCI support.
   3. Implement MDL-26226.
+  4. Course format API changes for M3.0.
 
 Version 2.9.1.1
   1. Update readme.